+++ styles/light/tag-sheet.mcss
@@ -1,0 +1,12 @@
+TagSheet {
+  padding: 20px
+
+  h2 {
+    font-weight: normal
+
+    input {
+      float: right
+      font-size: 100%
+    }
+  }
+}
